products/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cuser==2017.3.16
django-tinymce==4.1.0
progressbar2==4.5.0
pyserial==3.5
python-decouple==3.8
python-utils==3.9.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8
six==1.16.0